requirements.txt
PyGithub>=2.1.1
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0 